feat: first working version
It supports red, green, blue and combinations of them with max brightness (255) e.g. magenta (red+blue).

Leds can be turned on or off.

Operations on  LEDs are additive so if you want a specific colour regardless of the current state of the LEDs, you should switch all off with -o (or -w) followed by the colour you want.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>