fix: tweak pattern and add initialization delay
The controller's maximum time for a LED pattern as per the documentation is 5660 however that value doesn't work the first time the LED is turned on, possibly due to an initialization delay. 5653 seems to be working fine.

Similarly, we need a small delay between setting the pattern on hw_pattern and setting its duration on repeat, on the first instance the LED is initialized.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>