feat: leds-st1202: assume that patterns are indefinite if repeat is null
echoing patterns to hw_pattern followed by echoing a value to repeat complicates things in userspace.

There also seems to be a problem with the pattern trigger trying to start the pattern after storing the repeat value before the patterns are ready. This manifests in scripts where echoing values to hw_pattern and repeat occurs within milliseconds. See repeat_store in ledtrig-pattern.c for more reference. There is possibly need for a work queue somewhere.

So this change reduces complexity and seemingly avoids that issue when all we want is to turn leds on an off without taking care of repeat. Admittedly the  issue with repeat_store in ledtrig-pattern.c should be investigated in case something needs to be reported upstream.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>